% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mylake.R
\name{run_mylake}
\alias{run_mylake}
\title{Run the MyLake model}
\usage{
run_mylake(
  sim_folder = ".",
  config_dat = "mylake.Rdata",
  init_dat = "mylake_init.Rdata"
)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{config_dat}{the parameter file that needs to be run}

\item{init_dat}{the initial conditions file}
}
\description{
This runs the MyLake model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain a valid .Rdata file.
}
\author{
Jorrit Mesman, Tadhg Moore
}
\keyword{methods}
